Registrando e excluindo chaves Pix no Banco Central (BCB)

+++ src/main/kotlin/br/com/zupacademy/augusto/pix/deleta/DeletaPixService.kt
@@ -1,10 +1,16 @@
 package br.com.zupacademy.augusto.pix.deleta
 
+import br.com.zupacademy.augusto.client.DeletePixKeyRequest
+import br.com.zupacademy.augusto.client.Participant
+import br.com.zupacademy.augusto.client.SistemaPixBCBClient
 import br.com.zupacademy.augusto.pix.NotOwnerPixException
 import br.com.zupacademy.augusto.pix.PixRepository
+import io.micronaut.http.HttpStatus
+import io.micronaut.http.annotation.Part
 import io.micronaut.transaction.SynchronousTransactionManager
 import io.micronaut.validation.Validated
 import java.io.FileNotFoundException
+import java.lang.IllegalStateException
 import java.sql.Connection
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -14,7 +20,9 @@ import javax.validation.Valid
 @Singleton
 class DeletaPixService(
     @Inject val pixRepository: PixRepository,
-    @Inject val transactionManager: SynchronousTransactionManager<Connection>
+    @Inject val transactionManager: SynchronousTransactionManager<Connection>,
+    @Inject val bcbClient: SistemaPixBCBClient
+
 ) {
 
     fun deleta(@Valid deleteRequest: DeletaPixRequest) {
@@ -27,6 +35,12 @@ class DeletaPixService(
         if (pix.get().clienteId != deleteRequest.idCliente.toString()) {
             throw NotOwnerPixException("Somende o dono do pix pode excluí-la.")
         }
+
+        val request = DeletePixKeyRequest(pix.get().chave)
+
+        val bcbResponse = bcbClient.deleta(pix.get().chave, request)
+        if (bcbResponse.status != HttpStatus.OK)
+            throw IllegalStateException("Erro ao remover chave pix no BCB (Banco Central do Brasil).")
 
         transactionManager.executeWrite {
             pixRepository.delete(pix.get())
